chore: rename and restructure components
