+++ src/lib/components/News.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import ArticleLink from '$lib/components/article/link.svelte';
+	import ArticleLink from '$lib/components/ArticleLink.svelte';
 	import { format, parseISO } from 'date-fns/index.js';
 	import { nb } from 'date-fns/locale/index.js';
 	import type { Article } from '$lib/types/b3';
