--- src/lib/components/hero.svelte
@@ -1,6 +1,0 @@
-<script lang="ts"></script>
-
-<div
-	class="w-full aspect-video relative overflow-hidden h-96 xl:h-112 bg-cover"
-	style="background-image: url('/cover.webp')"
-></div>
