+++ src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script>
 	import '../app.css';
-	import Footer from '$lib/components/footer.svelte';
+	import Footer from '$lib/components/Footer.svelte';
 </script>
 
 <slot />
