+++ src/routes/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-	import Header from '$lib/components/header.svelte';
-	import Card from '$lib/components/card.svelte';
-	import News from '$lib/components/news.svelte';
-	import HorizontalPageDivider from '$lib/components/horizontalPageDivider.svelte';
+	import Header from '$lib/components/Header.svelte';
+	import Card from '$lib/components/Card.svelte';
+	import News from '$lib/components/News.svelte';
+	import HorizontalPageDivider from '$lib/components/PageDivider.svelte';
 	import type { HomePageData } from '$lib/types/b3.pagedata';
 	export let data: HomePageData;
 	const { topBanner, menuItems, articles } = data;
